Added ability to override flags field and push strings
The 'flags' field may be specified in the config object passed to
Open(). It should be an integer representing the read/write mode to use
when opening the queue (different from creating).

Additionally, a string may also be passed to Send() in addition to a
Buffer object for the user's convenience.